forex-bias-bot/requirements.txt
groq>=0.4.0
newsapi-python>=0.1.3
yfinance>=0.2.0
ta>=0.10.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
requests>=2.31.0
